refactored to split storage from main classes
